Use groupbox for main UI
* Add widget location info
