+++ MobileWeather/CurrentConditionsView.swift
@@ -15,40 +15,41 @@ struct CurrentConditionsView: View {
 #if DEBUG
         let _ = Self._printChanges()
 #endif
+        GroupBox {
+            VStack(alignment: .center, spacing: 10) {
+                HStack(alignment: .center, spacing: 10) {
+                    let iconName = OpenWeatherIcon(rawValue: currentForecast.conditionIconNames.first ?? OpenWeatherIcon.clearDay.rawValue)!
+                    WeatherImageView(openWeatherName: iconName)
+                        .frame(width: 100, height: 100)
 
-        VStack(alignment: .center, spacing: 10) {
-            HStack(alignment: .center, spacing: 10) {
-                let iconName = OpenWeatherIcon(rawValue: currentForecast.conditionIconNames.first ?? OpenWeatherIcon.clearDay.rawValue)!
-                WeatherImageView(openWeatherName: iconName)
-                    .frame(width: 100, height: 100)
+                    Spacer()
 
-                Spacer()
-
-                VStack(alignment: .trailing, spacing: 5) {
-                    Text(WeatherFormatter.temperatureFormatter.string(from: currentForecast.temperature))
-                        .font(.title)
-                        .fontWeight(.bold)
-                    HStack {
-                        Image(systemName: "wind")
-                        Text(currentForecast.windSpeed.formatted())
+                    VStack(alignment: .trailing, spacing: 5) {
+                        Text(WeatherFormatter.temperatureFormatter.string(from: currentForecast.temperature))
+                            .font(.title)
+                            .fontWeight(.bold)
+                        HStack {
+                            Image(systemName: "wind")
+                            Text(currentForecast.windSpeed.formatted())
+                        }
+                        HStack {
+                            Image(systemName: "sunrise")
+                            Text(currentForecast.sunriseDate.formatted(date: .omitted, time: .shortened))
+                        }
+                        HStack {
+                            Image(systemName: "sunset")
+                            Text(currentForecast.sunsetDate.formatted(date: .omitted, time: .shortened))
+                        }
+                        Text("UV Index: \(Int(currentForecast.uvIndex.rounded()))")
                     }
-                    HStack {
-                        Image(systemName: "sunrise")
-                        Text(currentForecast.sunriseDate.formatted(date: .omitted, time: .shortened))
-                    }
-                    HStack {
-                        Image(systemName: "sunset")
-                        Text(currentForecast.sunsetDate.formatted(date: .omitted, time: .shortened))
-                    }
-                    Text("UV Index: \(Int(currentForecast.uvIndex.rounded()))")
                 }
-            }
 
-            Text("\(currentForecast.conditionDescriptions.joined(separator: ", ").capitalized)")
-                .font(.title2)
-                .fontWeight(.medium)
+                Text("\(currentForecast.conditionDescriptions.joined(separator: ", ").capitalized)")
+                    .font(.title2)
+                    .fontWeight(.medium)
+            }
+            .padding([.horizontal], 40)
         }
-        .padding([.horizontal], 40)
     }
 }
