+++ MobileWeather/DailyConditionsView.swift
@@ -10,16 +10,17 @@ import SwiftUI
 
 struct DailyConditionsView: View {
     var dailyForecast: [DailyForecast]
-
+    
     var body: some View {
-        Text("Daily")
-            .font(.title)
-            .fontWeight(.bold)
-
-        VStack(alignment: .center, spacing: 20) {
-            ForEach(dailyForecast) { forecast in
-                DayForecastView(forecast: forecast)
+        GroupBox {
+            VStack(alignment: .center, spacing: 20) {
+                ForEach(dailyForecast) { forecast in
+                    DayForecastView(forecast: forecast)
+                }
             }
+        } label: {
+            Text("Daily")
+                .font(.title).fontWeight(.bold)
         }
     }
 }
@@ -29,7 +30,7 @@ struct DailyConditionsView_Previews: PreviewProvider {
         let json = Bundle.main.url(forResource: "weather-api-response", withExtension: "json")!
         let jsonData = try! Data(contentsOf: json)
         let data = try! JSONDecoder().decode(WeatherData.self, from: jsonData)
-
+        
         DailyConditionsView(dailyForecast: data.daily)
     }
 }
