+++ Podfile
@@ -1,4 +1,4 @@
-platform :ios, '13.0'
+platform :ios, '15.0'
 use_frameworks!
 
 target 'MobileWeather' do
