+++ SDL Weather Widget/UI/Assets.xcassets/AccentColor.colorset/Contents.json
