+++ SDL Weather Widget/UI/Assets.xcassets/AppIcon.appiconset/Contents.json
