+++ SDL Weather Widget/UI/Assets.xcassets/Contents.json
