+++ SDL Weather Widget/UI/Assets.xcassets/WidgetBackground.colorset/Contents.json
